\name{BioAssays-class}
\docType{class}
\alias{BioAssays-class}

\title{Class "BioAssays"}
\description{Abstract base class to represent a collection of
  "BioAssayData" objects}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  No methods defined with class "BioAssays" in the signature.
}
\author{Martin Morgan <mtmorgan@fhcrc.org>}
\seealso{}
\keyword{classes}
