\name{DerivedBioAssays-class}
\docType{class}
\alias{DerivedBioAssays-class}

\title{Class "DerivedBioAssays"}
\description{A class representing a collection of derived expression values}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("DerivedBioAssays", ...)}.
}
\section{Extends}{
  Class \code{"BioAssays"}, directly.
}
\section{Methods}{
  No methods defined with class "DerivedBioAssays" in the signature.
}
\author{Martin Morgan <mtmorgan@fhcrc.org>}
\seealso{}
\keyword{classes}
