\name{MeasuredBioAssayLocal-class}
\docType{class}
\alias{MeasuredBioAssayLocal-class}

\title{Class "MeasuredBioAssayLocal"}
\description{A class representing a collection of measured expression
  values, with expression values stored as a character vector of
  locally-accesible CEL files.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("MeasuredBioAssayLocal", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{bioAssayData}:}{Vector of CEL file names. Object of class \code{"character"}.}
  }
}
\section{Extends}{
  Class \code{"MeasuredBioAssayDataCollection"}, directly.
  Class \code{"BioAssayDataCollection"}, by class "MeasuredBioAssayDataCollection", distance 2.
}
\section{Methods}{
  No methods defined with class "MeasuredBioAssayLocal" in the signature.
}
\author{Martin Morgan <mtmorgan@fhcrc.org>}
\seealso{}
\keyword{classes}
