\name{MeasuredBioAssays-class}
\docType{class}
\alias{MeasuredBioAssays-class}

\title{Class "MeasuredBioAssays"}
\description{A class representing a collection of measured expression values}
\section{Objects from the Class}{
  Objects can be created by calls of the form \code{new("MeasuredBioAssays", ...)}.
}
\section{Extends}{
  Class \code{"BioAssays"}, directly.
}
\section{Methods}{
  No methods defined with class "MeasuredBioAssays" in the signature.
}
\author{Martin Morgan <mtmorgan@fhcrc.org>}
\seealso{}
\keyword{classes}
